Create rasterview
RasterView (macOS name) or rasterview (Linux name) is a small GUI application spezialized to visualize CUPS-, PWG- and Apple-raster files as they are generated by CUPS and other software to print to AirPrint capable or IPP Everywhere compliant printers.

It allows you to look at the (multipage) raster data produced for printing by any of the standard CUPS RIP filters (cgpdftoraster, imagetoraster, pdftoraster, and pstoraster). It is normally used to either test those filters or visualize at the data that is being sent to your raster printer driver. Additionally it is able to display the 'Page Attributes' which are embedded into the data to control the behavior of printer drivers

RasterView is free software provided under version 2 of the GNU General Public license.