Clean up worker.sh
Sorry about so many changes at once, but this really needs cleaning up.  

I added a simple exit function to try and make exit statuses more clear.  

I changed a few hard to follow if checks into cases.  

I tried to go through and make sure that all of the if statements are clean and that everything is quoted properly.  

I got rid of the remaining if statements using `*` where it would not work.

I added `set -v` before the huge for loop and `set +v` directly after.