Allow providing a webpage: attribute
Providing a link to the official webpage from where AppImages can be downloaded would be great additional information to the end-user. This would also enable web crawlers to automatically parse the AppImage link from the website instead of an absolute path. This would be helpful for AppImages that do not have their codebase on GitHub like Krita and Inkscape.