Add LBRY-desktop (Odysee electron-based player)
Please review as I've not contributed before. I looked at existing projects and some of them link to a specific release, whereas others link to the main github repo. I chose the latter because it is likely more "stable" (in that I hope it does not require editing the file on every new release).

As far as I can tell, the releases page of LBRY-desktop lists the most recent at the top, which I think is what is required for this to work properly. The releases also contain windows/debian/rpm packages, hopefully the catalogue is smart enough to detect the ".AppImage" file?

Let me know if I need to point to a specific release instead and updated it as new releases appear...

Also, how do I test this? Does it need to be merged or can I test with something like AppImagePool and some special config to point to my fork?