fix 'INVALID MESSAGE WITH QUOTES' screenshot issue, create Download button (#3226)
* Update lux

The latest upload has an AppImage of lux 1.1.8 built on ubuntu 20.04 instead of 18.04. It also has a new version of metadata.xml

* added code to produce a download link for AppImages in bitbucket repos with download section

* removed two misplaced comments which stopped the Download link from
entering the .md file

* committed data/lux to trigger new catalog entry rendition:
this version should provide a better default screenshot, avoiding the
INVALID MESSAGE WITH QUOTES output. With the recent changes in worker.sh,
re-rendering the catalog entry should also produce a 'download' button.

* Trigger re-run

* Trigger re-run

* Trigger test re-run

* Trigger re-run

* Trigger test re-run

* Trigger re-run

---------

Co-authored-by: Kay F. Jahnke <kfjahnke+pv@gmail.com>
Co-authored-by: probonopd <probonopd@users.noreply.github.com>